requirements.txt
langchain
langchain-groq
langchain-core
streamlit
pydantic
langgraph
python-dotenv
langchain-community
ddgs